Backend

+++ backend/src/Inmobiliaria.Application/Customers/Delete/DeleteCustomerHandler.cs
@@ -1,17 +1,19 @@
 using Inmobiliaria.Application.Shared;
 using Inmobiliaria.Domain.Customers;
-using Inmobiliaria.Domain.Shared.Exceptions;
-using Inmobiliaria.Domain.Shared.Resources;
 using MediatR;
 
 namespace Inmobiliaria.Application.Customers.Delete;
 
-public class DeleteCustomerHandler(IMapper mapper, ICustomerRepository customerRepository) : IRequestHandler<DeleteCustomerCommand, DeletedCustomerResult>
+public class DeleteCustomerHandler(IMapper mapper, ICustomerRepository customerRepository) : IRequestHandler<DeleteCustomerCommand, DeletedCustomerResult?>
 {
-    public async Task<DeletedCustomerResult> Handle(DeleteCustomerCommand request, CancellationToken cancellationToken)
+    public async Task<DeletedCustomerResult?> Handle(DeleteCustomerCommand request, CancellationToken cancellationToken)
     {
-        Customer customer = await customerRepository.GetByIdAsync(request.Id, cancellationToken)
-            ?? throw new DomainException(Errores.ResourceNotFound, nameof(customer));
+        Customer? customer = await customerRepository.GetByIdAsync(request.Id, cancellationToken);
+
+        if (customer is null)
+        {
+            return null;
+        }
 
         await customerRepository.DeleteAsync(customer, cancellationToken);
         DeletedCustomerResult result = mapper.ToDeletedCustomer(customer);
